This PR improves FlashMLA performance on the CPU for token generation (TG) with long contexts. The same strategy should also improve FA performance of GQA models, but something is not quite right there, so I have enabled only for MLA for now.

Here is a performance comparison between the main branch and this PR for DeepSeek-Lite on a Ryzen-7950X CPU

| model                |          test |    t/s (main)    |      t/s (PR)    |  Speedup |
| ---------------------| ------------: | ---------------: | ---------------: | -------: |
| deepseek2 16B IQ4_NL |    tg64@pp128 |     32.41 ± 0.04 |     32.22 ± 0.02 |  0.994   |
| deepseek2 16B IQ4_NL |    tg64@pp256 |     32.16 ± 0.02 |     31.96 ± 0.03 |  0.994   |
| deepseek2 16B IQ4_NL |    tg64@pp512 |     31.80 ± 0.00 |     31.85 ± 0.05 |  1.002   |
| deepseek2 16B IQ4_NL |   tg64@pp1024 |     31.30 ± 0.03 |     31.51 ± 0.00 |  1.007   |
| deepseek2 16B IQ4_NL |   tg64@pp2048 |     30.44 ± 0.01 |     30.93 ± 0.02 |  1.016   |
| deepseek2 16B IQ4_NL |   tg64@pp4096 |     28.50 ± 0.01 |     29.69 ± 0.08 |  1.042   |
| deepseek2 16B IQ4_NL |   tg64@pp8192 |     25.31 ± 0.14 |     27.19 ± 0.11 |  1.074   |
| deepseek2 16B IQ4_NL |  tg64@pp16384 |     20.40 ± 0.10 |     22.31 ± 0.03 |  1.094   |

For TG the `V*softmax(K*Q)` is parallelized along the heads, so given enough threads, the `K*Q` operation computed by each thread becomes a GEMV, which is notoriously memory bound. In this PR parallelization is done along the K-cache entries, with the `K*Q` portions computed by each thread being GEMM, which is faster. But this requires one additional thread synchronization before combining the results of the threads. My guess is that this extra barrier leads to the observed slightly lower performance for short contexts (where with the main branch implementation `K*Q` is fast despite being GEMV).

To put the above table into perspective, TG speed with a context of 16k tokens is around 10 t/s without MLA and FA for this model on this CPU.