@ikawrakow 2025-03-07
The above table is for `Q8_KV` KV cache. Here is a comparison between the main branch and this PR for `fp16` KV cache:

| model                |          test |      t/s (main)  |      t/s (PR)    |  Speedup |
| ---------------------| ------------: | ---------------: | ---------------: | -------: |
| deepseek2 16B IQ4_NL |    tg64@pp128 |     31.54 ± 0.06 |     32.24 ± 0.01 |  1.022   |
| deepseek2 16B IQ4_NL |    tg64@pp256 |     30.79 ± 0.08 |     31.86 ± 0.05 |  1.035   |
| deepseek2 16B IQ4_NL |    tg64@pp512 |     29.83 ± 0.02 |     31.90 ± 0.01 |  1.069   |
| deepseek2 16B IQ4_NL |   tg64@pp1024 |     28.48 ± 0.02 |     31.48 ± 0.03 |  1.105   |
| deepseek2 16B IQ4_NL |   tg64@pp2048 |     26.05 ± 0.01 |     30.69 ± 0.00 |  1.178   |
| deepseek2 16B IQ4_NL |   tg64@pp4096 |     22.12 ± 0.04 |     29.45 ± 0.05 |  1.331   |
| deepseek2 16B IQ4_NL |   tg64@pp8192 |     17.25 ± 0.16 |     27.37 ± 0.14 |  1.587   |
| deepseek2 16B IQ4_NL |  tg64@pp16384 |     11.78 ± 0.03 |     23.13 ± 0.64 |  1.963   |

I.e., the PR is a massive upgrade in this case (but it also tells us that the original `fp16` FA kernel was far from optimal).